Move communication to new gamelogic android library
