--- communication/src/main/java/at/aau/se2/communication/models/GameState.java
@@ -1,3 +1,0 @@
-package at.aau.se2.communication.models;
-
-public enum GameState {}
