+++ gamelogic/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'java-library'
+    id 'com.android.library'
 }
 
 apply plugin: 'com.diffplug.spotless'
@@ -13,14 +13,33 @@ spotless {
     }
 }
 
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+android {
+    compileSdk 32
+
+    defaultConfig {
+        minSdk 23
+        targetSdk 32
+
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    implementation 'junit:junit:4.12'
-    implementation 'androidx.annotation:annotation:1.2.0'
     testImplementation "org.mockito:mockito-core:3.+"
+    testImplementation 'junit:junit:4.13.2'
 
+    implementation 'androidx.annotation:annotation:1.2.0'
+    implementation platform('com.google.firebase:firebase-bom:29.3.0')
+    implementation 'com.google.firebase:firebase-database'
 }
