+++ gamelogic/src/main/java/at/aau/se2/gamelogic/comunication/ResultObserver.java
@@ -1,4 +1,4 @@
-package at.aau.se2.communication;
+package at.aau.se2.gamelogic.comunication;
 
 public interface ResultObserver<T, U> {
   void finished(Result<T, U> result);
